Refactors in to own file. Now inherits from LxBrNet
